employee

+++ database/migrations/2025_03_01_022203_create_leave_requests_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('leave_requests', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('employee_id'); 
+            $table->unsignedBigInteger('employee_id');
             $table->foreign('employee_id')->references('id')->on('employees')->onDelete('cascade');
             $table->enum('leave_type', ['sick', 'vacation', 'casual']);
             $table->date('start_date');
